+++ locales/cy/aml-body-number.json
@@ -7,5 +7,6 @@
   "amlcontactinfo": "Cysylltwch â'r corff goruchwylio AML os oes angen i chi wirio'ch manylion.",
 
   "error-amlIDNumberInput": "Cofnodwch y manylion ar gyfer ",
-  "error-amlIdLength": "Rhaid i'r rhif aelodaeth Gwrth-Gwyngalchu Arian (AML) fod yn 256 nod neu lai ar gyfer "
+  "error-amlIdLength": "Rhaid i'r rhif aelodaeth Gwrth-Gwyngalchu Arian (AML) fod yn 256 nod neu lai ar gyfer ",
+  "error-duplicatedAmlMembership": "The membership number you entered has already been added for this AML supervisory body. Enter a different membership number. Welsh"
 }
